+++ matrix-sdk/src/main/java/org/matrix/androidsdk/data/Room.java
@@ -1940,7 +1940,7 @@ public class Room {
             }
 
             if (null != getStore()) {
-                getStore().storeAccountData(getRoomId(), mAccountData);
+                getStore().storeRoomAccountData(getRoomId(), mAccountData);
             }
         }
     }